+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Copper's stats
+### Fixed
+- Adamant, Aquatic, and Chilled traits not having their bonus damage effect [#373]
+
 ## [2.6.28] - 2021-07-03
 ### Added
 - Enchantment description for Star Charged (TomasBorsje) [#399, #388]